Use Neon submodule's dependencies
This should help us stay up to date in the future. The setup-neon script
performs the same actions as the build action in the previous
package.json. Unneeded files are removed. The webpack configuration had
to be altered as now the submodule is the working directory, not the
wrapper root.